requirements.txt
fastapi
uvicorn
python-dotenv>=1.0.0
httpx>=0.27.0
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.0.0
yfinance>=0.2.0